+++ packages/web-component/src/lib/descope-wc/BaseDescopeWc.ts
@@ -35,6 +35,10 @@ import {
   FlowConfig,
 } from '../types';
 import initTemplate from './initTemplate';
+import {
+  extractNestedAttribute,
+  transformFlowInputFormData,
+} from '../helpers/flowInputs';
 
 // this is replaced in build time
 declare const BUILD_VERSION: string;
@@ -70,6 +74,10 @@ class BaseDescopeWc extends HTMLElement {
 
   #init = false;
 
+  formConfig = transformFlowInputFormData(this.getAttribute('form'));
+
+  formConfigValues = extractNestedAttribute(this.formConfig, 'value');
+
   loggerWrapper = {
     error: (message: string, description = '') => {
       this.logger.error(message, description, new Error());
@@ -80,6 +88,9 @@ class BaseDescopeWc extends HTMLElement {
     },
     info: (message: string, description = '', state: any = {}) => {
       this.logger.info(message, description, state);
+    },
+    debug: (message: string, description = '') => {
+      this.logger.debug(message, description);
     },
   };
 
@@ -147,24 +158,6 @@ class BaseDescopeWc extends HTMLElement {
 
   get flowId() {
     return this.getAttribute('flow-id');
-  }
-
-  get form() {
-    try {
-      const form = (JSON.parse(this.getAttribute('form')) || {}) as Record<
-        string,
-        any
-      >;
-      return Object.entries(form).reduce(
-        (prev, [key, value]) => ({
-          ...prev,
-          [`form.${key}`]: value,
-        }),
-        form,
-      );
-    } catch (e) {
-      return {};
-    }
   }
 
   get client() {
@@ -399,7 +392,10 @@ class BaseDescopeWc extends HTMLElement {
 
     const descopeUi = await BaseDescopeWc.descopeUI;
 
-    if (descopeUi?.componentsThemeManager) {
+    if (
+      descopeUi?.componentsThemeManager &&
+      !descopeUi.componentsThemeManager.hasThemes
+    ) {
       descopeUi.componentsThemeManager.themes = {
         light: theme?.light?.components,
         dark: theme?.dark?.components,
@@ -557,7 +553,7 @@ class BaseDescopeWc extends HTMLElement {
 
   async #loadDescopeUI() {
     if (BaseDescopeWc.descopeUI) {
-      this.loggerWrapper.info(
+      this.loggerWrapper.debug(
         'DescopeUI is already loading, probably multiple flows are running on the same page',
       );
       return;